(mionor) make minor corrections - rename state and method name, char limit value, remove return statement
